+++ src/CouchDB.Driver/CouchClientAuthentication.cs
@@ -1,8 +1,11 @@
-﻿using CouchDB.Driver.Settings;
+﻿using CouchDB.Driver.DTOs;
+using CouchDB.Driver.Exceptions;
+using CouchDB.Driver.Settings;
 using Flurl.Http;
 using Nito.AsyncEx;
 using System;
 using System.Linq;
+using System.Net.Http;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
@@ -10,10 +13,10 @@ namespace CouchDB.Driver
 {
     public partial class CouchClient
     {
-        protected virtual void OnBeforeCall(HttpCall call)
+        protected virtual void OnBeforeCall(HttpCall httpCall)
         {
             // If session requests no authorization needed
-            if (call.Request.RequestUri.ToString().Contains("_session"))
+            if (httpCall.Request.RequestUri.ToString().Contains("_session"))
             {
                 return;
             }
@@ -22,7 +25,7 @@ namespace CouchDB.Driver
                 case AuthenticationType.None:
                     break;
                 case AuthenticationType.Basic:
-                    call.FlurlRequest.WithBasicAuth(_settings.Username, _settings.Password);
+                    httpCall.FlurlRequest.WithBasicAuth(_settings.Username, _settings.Password);
                     break;
                 case AuthenticationType.Cookie:
                     var isTokenExpired =
@@ -32,7 +35,7 @@ namespace CouchDB.Driver
                     {
                         AsyncContext.Run(() => LoginAsync());
                     }
-                    call.FlurlRequest.EnableCookies().WithCookie("AuthSession", _cookieToken);
+                    httpCall.FlurlRequest.EnableCookies().WithCookie("AuthSession", _cookieToken);
                     break;
                 default:
                     throw new NotSupportedException($"Authentication of type {_settings.AuthenticationType} is not supported.");
@@ -41,13 +44,14 @@ namespace CouchDB.Driver
 
         private async Task LoginAsync()
         {
-            var response = await _flurlClient.Request(ConnectionString)
+            HttpResponseMessage response = await _flurlClient.Request(ConnectionString)
                 .AppendPathSegment("_session")
                 .PostJsonAsync(new
                 {
                     name = _settings.Username,
                     password = _settings.Password
-                });
+                })
+                .ConfigureAwait(false);
 
             _cookieCreationDate = DateTime.Now;
 
@@ -55,7 +59,7 @@ namespace CouchDB.Driver
             {
                 var dirtyToken = values.First();
                 var regex = new Regex(@"^AuthSession=(.+); Version=1; .*Path=\/; HttpOnly$");
-                var match = regex.Match(dirtyToken);
+                Match match = regex.Match(dirtyToken);
                 if (match.Success)
                 {
                     _cookieToken = match.Groups[1].Value;
@@ -67,9 +71,16 @@ namespace CouchDB.Driver
         }
         private async Task LogoutAsync()
         {
-            await _flurlClient.Request(ConnectionString)
+            OperationResult result = await _flurlClient.Request(ConnectionString)
                 .AppendPathSegment("_session")
-                .DeleteAsync();
+                .DeleteAsync()
+                .ReceiveJson<OperationResult>()
+                .ConfigureAwait(false);
+
+            if (!result.Ok)
+            {
+                throw new CouchDeleteException();
+            }
         }
     }
 }