+++ src/CouchDB.Driver/DTOs/FindResult.cs
@@ -4,7 +4,9 @@ using System.Collections.Generic;
 
 namespace CouchDB.Driver.DTOs
 {
+#pragma warning disable CA1812 // Avoid uninstantiated internal classes
     internal class FindResult<T>
+#pragma warning restore CA1812 // Avoid uninstantiated internal classes
     {
         [JsonProperty("docs")]
         public IEnumerable<T> Docs { get; internal set; }