+++ src/CouchDB.Driver/Extensions/EnumerableExtensions.cs
@@ -13,12 +13,16 @@ namespace CouchDB.Driver.Extensions
         /// <param name="source">A sequence in which to locate a value.</param>
         /// <param name="input">Values to locate in the sequence.</param>
         /// <returns>true if the source sequence contains all elements that has specified values; otherwise, false.</returns>
-        public static bool Contains<T>(this IEnumerable<T> source, IEnumerable<T> input) where T : IConvertible
+        public static bool Contains<T>(this IEnumerable<T> source, IEnumerable<T> input)
         {
             if (source == null)
+            {
                 throw new ArgumentNullException(nameof(source));
+            }
             if (input == null)
+            {
                 throw new ArgumentNullException(nameof(input));
+            }
 
             return input.All(s => source.Contains(s));
         }