+++ src/CouchDB.Driver/Types/CouchType.cs
@@ -18,26 +18,26 @@
         /// <summary>
         /// Represents the "null" type.
         /// </summary>
-        public static readonly CouchType Null = new CouchType("null");
+        public static readonly CouchType CNull = new CouchType("null");
         /// <summary>
         /// Represents the "boolean" type.
         /// </summary>
-        public static readonly CouchType Boolean = new CouchType("boolean");
+        public static readonly CouchType CBoolean = new CouchType("boolean");
         /// <summary>
         /// Represents the "number" type.
         /// </summary>
-        public static readonly CouchType Number = new CouchType("number");
+        public static readonly CouchType CNumber = new CouchType("number");
         /// <summary>
         /// Represents the "string" type.
         /// </summary>
-        public static readonly CouchType String = new CouchType("string");
+        public static readonly CouchType CString = new CouchType("string");
         /// <summary>
         /// Represents the "array" type.
         /// </summary>
-        public static readonly CouchType Array = new CouchType("array");
+        public static readonly CouchType CArray = new CouchType("array");
         /// <summary>
         /// Represents the "object" type.
         /// </summary>
-        public static readonly CouchType Object = new CouchType("object");
+        public static readonly CouchType CObject = new CouchType("object");
     }
 }