+++ tests/CouchDB.Driver.UnitTests/Database_Tests.cs
@@ -3,6 +3,8 @@ using CouchDB.Driver.Types;
 using CouchDB.Driver.UnitTests.Models;
 using Flurl.Http.Testing;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Xunit;
@@ -26,6 +28,7 @@ namespace CouchDB.Driver.UnitTests
         {
             using (var httpTest = new HttpTest())
             {
+                var a = new List<Rebel>();
                 var newR = await _rebels.FindAsync("1");
                 httpTest
                     .ShouldHaveCalled("http://localhost/rebels/1")
@@ -43,6 +46,31 @@ namespace CouchDB.Driver.UnitTests
                     .ShouldHaveCalled("http://localhost/rebels/1")
                     .WithQueryParamValue("conflicts", true)
                     .WithVerb(HttpMethod.Get);
+            }
+        }
+        [Fact]
+        public async Task FindMany()
+        {
+            using (var httpTest = new HttpTest())
+            {
+                httpTest.RespondWith(@"{""results"":[{""id"":""1"",""docs"":[{""ok"":{""_id"":""1"",""Name"":""Luke""}}]},{""id"":""2"",""docs"":[{""ok"":{""_id"":""2"",""Name"":""Leia""}}]}]}");
+                var ids = new string[] { "1", "2" };
+                var result = await _rebels.FindManyAsync(ids);
+                httpTest
+                    .ShouldHaveCalled("http://localhost/rebels/_bulk_get")
+                    .WithRequestJson(new
+                    {
+                        docs = new[]
+                        {
+                            new { id = "1" },
+                            new { id = "2" },
+                        }
+                    })
+                    .WithVerb(HttpMethod.Post);
+
+                Assert.Equal(2, result.Count);
+                Assert.Equal("Luke", result[0].Name);
+                Assert.Equal("Leia", result[1].Name);
             }
         }
         [Fact]
@@ -92,6 +120,9 @@ namespace CouchDB.Driver.UnitTests
         {
             using (var httpTest = new HttpTest())
             {
+                // Operation response
+                httpTest.RespondWithJson(new { ok = true });
+
                 var r = new Rebel { Name = "Luke", Id = "1", Rev = "1" };
                 await _rebels.DeleteAsync(r);
                 httpTest
@@ -104,7 +135,10 @@ namespace CouchDB.Driver.UnitTests
         {
             using (var httpTest = new HttpTest())
             {
-                httpTest.RespondWithJson(new { Docs = new string[0], Bookmark = "bookmark" });
+                // ToList
+                httpTest.RespondWithJson(new { Docs = new List<string>(), Bookmark = "bookmark" });
+                // Operation response
+                httpTest.RespondWithJson(new { ok = true });
 
                 using (var client = new CouchClient("http://localhost"))
                 {
@@ -119,6 +153,43 @@ namespace CouchDB.Driver.UnitTests
             }
         }
 
+        [Fact]
+        public async Task QueryJson()
+        {
+            using (var httpTest = new HttpTest())
+            {
+                var expected = new List<Rebel>() { new Rebel { Id = Guid.NewGuid().ToString() } };
+                httpTest.RespondWithJson(new { Docs = expected });
+
+                var query = @"{""selector"":{""age"":19}}";
+                var result = await _rebels.QueryAsync(query);
+                httpTest
+                    .ShouldHaveCalled("http://localhost/rebels/_find")
+                    .WithVerb(HttpMethod.Post)
+                    .WithRequestBody(@"{""selector"":{""age"":19}}");
+                Assert.Equal(expected.Count, result.Count);
+                Assert.Equal(expected[0].Id, result[0].Id);
+            }
+        }
+        [Fact]
+        public async Task QueryObject()
+        {
+            using (var httpTest = new HttpTest())
+            {
+                var expected = new List<Rebel>() { new Rebel { Id = Guid.NewGuid().ToString() } };
+                httpTest.RespondWithJson(new { Docs = expected });
+
+                var query = new { selector = new { age = 19 } };
+                var result = await _rebels.QueryAsync(query);
+                httpTest
+                    .ShouldHaveCalled("http://localhost/rebels/_find")
+                    .WithVerb(HttpMethod.Post)
+                    .WithRequestBody(@"{""selector"":{""age"":19}}");
+                Assert.Equal(expected.Count, result.Count);
+                Assert.Equal(expected[0].Id, result[0].Id);
+            }
+        }
+
         #endregion
 
         #region Bulk
@@ -128,10 +199,13 @@ namespace CouchDB.Driver.UnitTests
         {
             using (var httpTest = new HttpTest())
             {
+                // Response
                 httpTest.RespondWithJson(new[] {
                     new { Id = "111", Ok = true, Rev = "111" },
                     new { Id = "222", Ok = true, Rev = "222" },
                 });
+                // Logout
+                httpTest.RespondWithJson(new { ok = true });
 
                 var moreRebels = new[] {
                     new Rebel { Name = "Luke", Id = "1" },
@@ -164,6 +238,9 @@ namespace CouchDB.Driver.UnitTests
         {
             using (var httpTest = new HttpTest())
             {
+                // Operation response
+                httpTest.RespondWithJson(new { ok = true });
+
                 await _rebels.CompactAsync();
                 httpTest
                     .ShouldHaveCalled("http://localhost/rebels/_compact")